Played around with charts
No pub sub yet.  Just played around with how charts work.  Determined
it was best to extend each type needed with a function that
encapsulates the data change and update.  Created a function to do this
and create charts for each type listed in an array.  Will need to add
some unit testing from here on out to ensure the publisher is built
correctly.  Also need to backfill tests for my chart creation function.